+++ packages/react-impulse-form/tests/impulse-form-unit/impulse-form-unit.spec.ts
@@ -1,0 +1,380 @@
+import { z } from "zod"
+
+import { ImpulseFormUnit } from "../../src"
+
+it("creates ImpulseFormUnit without validation", ({ scope }) => {
+  const value = ImpulseFormUnit(1)
+
+  expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<number>>()
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-arguments
+  expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<number, null>>()
+  expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<number, null, number>>()
+
+  expect(value.getInput(scope)).toBe(1)
+  expect(value.getOutput(scope)).toBe(1)
+})
+
+it("creates ImpulseFormUnit with same type validator", ({ scope }) => {
+  const value = ImpulseFormUnit("", {
+    validateOn: "onInit",
+    validate: (input) => {
+      const trimmed = input.trim()
+
+      return trimmed.length < 2 ? ["too short", null] : [null, trimmed]
+    },
+  })
+
+  expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<string, string>>()
+  expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<string, string, string>>()
+
+  expect(value.getInput(scope)).toBe("")
+  expect(value.getOutput(scope)).toBeNull()
+
+  value.setInput(" 123 ")
+  expect(value.getInput(scope)).toBe(" 123 ")
+  expect(value.getOutput(scope)).toBe("123")
+})
+
+it("creates ImpulseFormUnit with same type schema", ({ scope }) => {
+  const value = ImpulseFormUnit("", {
+    schema: z.string().trim().min(2),
+    validateOn: "onInit",
+  })
+
+  expectTypeOf(value).toEqualTypeOf<
+    ImpulseFormUnit<string, ReadonlyArray<string>>
+  >()
+  expectTypeOf(value).toEqualTypeOf<
+    ImpulseFormUnit<string, ReadonlyArray<string>, string>
+  >()
+
+  expect(value.getInput(scope)).toBe("")
+  expect(value.getOutput(scope)).toBeNull()
+
+  value.setInput(" 123 ")
+  expect(value.getInput(scope)).toBe(" 123 ")
+  expect(value.getOutput(scope)).toBe("123")
+})
+
+it("creates ImpulseFormUnit with converting type schema", ({ scope }) => {
+  const value = ImpulseFormUnit("", {
+    schema: z.string().trim().min(1).pipe(z.coerce.number()),
+    validateOn: "onInit",
+  })
+
+  expectTypeOf(value).toEqualTypeOf<
+    ImpulseFormUnit<string, ReadonlyArray<string>, number>
+  >()
+
+  expect(value.getInput(scope)).toBe("")
+  expect(value.getOutput(scope)).toBeNull()
+
+  value.setInput(" 123 ")
+  expect(value.getInput(scope)).toBe(" 123 ")
+  expect(value.getOutput(scope)).toBe(123)
+})
+
+it("creates ImpulseFormUnit with complex value", ({ scope }) => {
+  const value = ImpulseFormUnit(
+    {
+      type: "",
+      value: "",
+    },
+    {
+      schema: z.object({
+        type: z.enum(["first", "second"]),
+        value: z.string().min(1).pipe(z.coerce.boolean()),
+      }),
+      validateOn: "onInit",
+    },
+  )
+  expectTypeOf(value).toEqualTypeOf<
+    ImpulseFormUnit<
+      {
+        type: string
+        value: string
+      },
+      ReadonlyArray<string>,
+      {
+        type: "first" | "second"
+        value: boolean
+      }
+    >
+  >()
+
+  expect(value.getInput(scope)).toStrictEqual({
+    type: "",
+    value: "",
+  })
+  expect(value.getOutput(scope)).toBeNull()
+
+  value.setInput((current) => ({ ...current, type: "first" }))
+  expect(value.getOutput(scope)).toBeNull()
+
+  value.setInput((current) => ({ ...current, value: "true" }))
+  expect(value.getOutput(scope)).toStrictEqual({
+    type: "first",
+    value: true,
+  })
+})
+
+it("does not allow to specify schema TOutput different from TInput", ({
+  scope,
+}) => {
+  const value = ImpulseFormUnit<string>("1", {
+    // @ts-expect-error schema Input is not string
+    schema: z.coerce.number(),
+    validateOn: "onInit",
+  })
+
+  expect(value.getInput(scope)).toBe("1")
+  expect(value.getOutput(scope)).toBe(1)
+})
+
+it("does not allow to specify schema Output different from TValue", ({
+  scope,
+}) => {
+  const value = ImpulseFormUnit<number, string>(0, {
+    // @ts-expect-error schema Output is not string
+    schema: z.number(),
+    validateOn: "onInit",
+  })
+
+  expect(value.getInput(scope)).toBe(0)
+  expect(value.getOutput(scope)).toBe(0)
+})
+
+it("specifies initial touched", ({ scope }) => {
+  const value = ImpulseFormUnit("", { touched: true })
+
+  expect(value.isTouched(scope)).toBe(true)
+
+  value.setTouched(false)
+  expect(value.isTouched(scope)).toBe(false)
+})
+
+it("specifies initial value", ({ scope }) => {
+  const value = ImpulseFormUnit("", { initial: "1" })
+
+  expect(value.getInput(scope)).toBe("")
+  expect(value.getInitial(scope)).toBe("1")
+})
+
+it("assigns initial to input by default", ({ scope }) => {
+  const value = ImpulseFormUnit("1")
+
+  expect(value.getInput(scope)).toBe("1")
+  expect(value.getInitial(scope)).toBe("1")
+})
+
+it("assigns custom initial", ({ scope }) => {
+  const value = ImpulseFormUnit("1", { initial: "2" })
+
+  expect(value.getInput(scope)).toBe("1")
+  expect(value.getInitial(scope)).toBe("2")
+})
+
+it("returns initial if it is equals to input with custom isInputEqual", ({
+  scope,
+}) => {
+  const initial = { count: 0 }
+  const form = ImpulseFormUnit(
+    { count: 0 },
+    {
+      initial: initial,
+      isInputEqual: (left, right) => left.count === right.count,
+    },
+  )
+
+  expect(form.getInput(scope)).toBe(initial)
+  expect(form.getInput(scope)).toBe(form.getInitial(scope))
+
+  form.setInput({ count: 1 })
+  expect(form.getInput(scope)).not.toBe(initial)
+})
+
+it("keeps the prev value with custom isInputEqual", ({ scope }) => {
+  const form = ImpulseFormUnit(
+    { count: 0 },
+    {
+      isInputEqual: (left, right) => left.count === right.count,
+    },
+  )
+
+  const input = form.getInput(scope)
+
+  form.setInput({ count: 0 })
+  expect(form.getInput(scope)).toBe(input)
+
+  form.setInput({ count: 1 })
+  expect(form.getInput(scope)).not.toBe(input)
+})
+
+describe("ImpulseFormUnitValidatedOptions", () => {
+  it("defines impulse with validate as transformer", ({ scope }) => {
+    const value = ImpulseFormUnit(123, {
+      validateOn: "onInit",
+      validate: (input) => [null, input.toFixed(2)],
+    })
+
+    expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<number, null, string>>()
+    expect(value.getInput(scope)).toBe(123)
+    expect(value.getOutput(scope)).toBe("123.00")
+  })
+
+  it("defines impulse with validate as validator", ({ scope }) => {
+    const value = ImpulseFormUnit(0, {
+      validateOn: "onInit",
+      validate: (input) => {
+        return input > 0 ? [null, input] : ["should be positive", null]
+      },
+    })
+
+    expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<number, string, number>>()
+    expect(value.getInput(scope)).toBe(0)
+    expect(value.getError(scope)).toBe("should be positive")
+    expect(value.getOutput(scope)).toBeNull()
+
+    value.setInput(1)
+    expect(value.getInput(scope)).toBe(1)
+    expect(value.getError(scope)).toBeNull()
+    expect(value.getOutput(scope)).toBe(1)
+  })
+
+  it("defines impulse with validate as validator and transformer", ({
+    scope,
+  }) => {
+    const value = ImpulseFormUnit(0, {
+      validateOn: "onInit",
+      validate: (input) => {
+        return input > 0
+          ? [null, input.toFixed(2)]
+          : ["should be positive", null]
+      },
+    })
+
+    expectTypeOf(value).toEqualTypeOf<ImpulseFormUnit<number, string, string>>()
+    expect(value.getInput(scope)).toBe(0)
+    expect(value.getError(scope)).toBe("should be positive")
+    expect(value.getOutput(scope)).toBeNull()
+
+    value.setInput(1)
+    expect(value.getInput(scope)).toBe(1)
+    expect(value.getError(scope)).toBeNull()
+    expect(value.getOutput(scope)).toBe("1.00")
+  })
+})
+
+describe("ImpulseFormUnitSchemaOptions", () => {
+  it("defines impulse with schema as validator", ({ scope }) => {
+    const value = ImpulseFormUnit(0, {
+      validateOn: "onInit",
+      schema: z.number().min(1),
+    })
+
+    expectTypeOf(value).toEqualTypeOf<
+      ImpulseFormUnit<number, ReadonlyArray<string>, number>
+    >()
+    expect(value.getInput(scope)).toBe(0)
+    expect(value.getError(scope)).toStrictEqual([
+      "Number must be greater than or equal to 1",
+    ])
+    expect(value.getOutput(scope)).toBeNull()
+
+    value.setInput(1)
+    expect(value.getInput(scope)).toBe(1)
+    expect(value.getError(scope)).toBeNull()
+    expect(value.getOutput(scope)).toBe(1)
+  })
+
+  it("defines impulse with schema as validator and transformer", ({
+    scope,
+  }) => {
+    const value = ImpulseFormUnit(0, {
+      validateOn: "onInit",
+      schema: z
+        .number()
+        .min(1)
+        .transform((input) => input.toFixed(2)),
+    })
+
+    expectTypeOf(value).toEqualTypeOf<
+      ImpulseFormUnit<number, ReadonlyArray<string>, string>
+    >()
+    expect(value.getInput(scope)).toBe(0)
+    expect(value.getError(scope)).toStrictEqual([
+      "Number must be greater than or equal to 1",
+    ])
+    expect(value.getOutput(scope)).toBeNull()
+
+    value.setInput(1)
+    expect(value.getInput(scope)).toBe(1)
+    expect(value.getError(scope)).toBeNull()
+    expect(value.getOutput(scope)).toBe("1.00")
+  })
+})
+
+describe("ImpulseFormUnitOptions.isInputDirty", () => {
+  it("uses Object.is when not provided", ({ scope }) => {
+    const value = ImpulseFormUnit({ value: "value" })
+
+    expect(value.isDirty(scope)).toBe(false)
+
+    value.setInput({ value: "value" })
+    expect(value.isDirty(scope)).toBe(true)
+  })
+
+  it("fallbacks to isInputEqual when not provided", ({ scope }) => {
+    const value = ImpulseFormUnit(
+      { value: "value" },
+      {
+        isInputEqual: (left, right) => left.value === right.value,
+      },
+    )
+
+    expect(value.isDirty(scope)).toBe(false)
+
+    value.setInput({ value: "value" })
+    expect(value.isDirty(scope)).toBe(false)
+  })
+
+  it("takes isInputDirty over isInputEqual", ({ scope }) => {
+    const value = ImpulseFormUnit(
+      { value: "value" },
+      {
+        isInputEqual: (left, right) => left.value === right.value,
+        isInputDirty: (left, right) => left.value.trim() !== right.value.trim(),
+      },
+    )
+
+    expect(value.isDirty(scope)).toBe(false)
+
+    value.setInput({ value: "value " })
+    expect(value.isDirty(scope)).toBe(false)
+
+    value.setInput({ value: "value 1" })
+    expect(value.isDirty(scope)).toBe(true)
+  })
+
+  it("returns not dirty when isInputDirty returns false", ({ scope }) => {
+    const value = ImpulseFormUnit("", {
+      isInputDirty: (left, right) => left.trim() !== right.trim(),
+    })
+
+    expect(value.isDirty(scope)).toBe(false)
+
+    value.setInput(" ")
+    expect(value.isDirty(scope)).toBe(false)
+  })
+
+  it("returns dirty when isInputDirty returns true", ({ scope }) => {
+    const value = ImpulseFormUnit("", {
+      isInputDirty: (left, right) => left.trim() !== right.trim(),
+    })
+
+    expect(value.isDirty(scope)).toBe(false)
+
+    value.setInput(" s")
+    expect(value.isDirty(scope)).toBe(true)
+  })
+})
